third_party/blink/web_tests/xr/exclusive_requestFrame_nolayer.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testdriver.js"></script>
<script src="../resources/testdriver-vendor.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas />

<script>
let immersiveTestName = "XRSession requestAnimationFrame must fail if the session has "
  + "no baseLayer for immersive";

let nonImmersiveTestName = "XRSession requestAnimationFrame must fail if the session has "
  + "no baseLayer for non immersive";

let fakeDeviceInitParams = TRACKED_IMMERSIVE_DEVICE;

let testFunction = (session, controller, t, sessionObjects) => new Promise((resolve, reject) => {
  // Clear the base layer that the boilerplate sets.  This ensures that the rAF
  // won't fire.  If the fire *does* fire, it will fail because the baseLayer
  // won't match the new baselayer we later create.
  session.updateRenderState({
    baseLayer: null
  });
  let webglCanvas = document.getElementsByTagName('canvas')[0];
  let gl = sessionObjects.gl;

  // Session must have a baseLayer or frame requests will be ignored.
  let webglLayer = new XRWebGLLayer(session, gl, {
      compositionDisabled: session.mode == 'inline' });

  function onEarlyFrame(time, vrFrame) {
    // We shouldn't be allowed to reach this callback with no baseLayer
    t.step(() => {
      assert_equals(session.renderState.baseLayer, webglLayer);
    });
    resolve();
  }

  // This callback shouldn't go through, since the session doesn't
  // have a baseLayer when this call is made.
  let handle = session.requestAnimationFrame(onEarlyFrame);
  // Should still give us a valid handle, though.
  assert_not_equals(handle, 0);

  // Wait for a bit and set the baseLayer.
  setTimeout(() => {
    // Once the base layer is set the previously registered callback should run.
    session.updateRenderState({
        baseLayer: webglLayer
    });
  }, 300);
});

xr_session_promise_test(
  immersiveTestName, testFunction, fakeDeviceInitParams, 'immersive-vr');
xr_session_promise_test(
  nonImmersiveTestName, testFunction, fakeDeviceInitParams, 'inline');

</script>
